add defaults for SidelineConfig, move Annotation class
